Moves util method to osgutils.py
